feat: update file format (tsx) and add args
